+++ fatiando/seismic/wavefd.py
@@ -1,6 +1,15 @@
 r"""
 Finite difference solution of the 2D wave equation for isotropic media.
 
+.. warning::
+
+    This code is experimental and poorly tested! We cannot guarantee that the
+    results are accurate. **We strongly discourage use of this code for
+    research purposes.** Still, the general behaviour of the waves seems
+    consistent with theory and can be useful for illustrating how seismic waves
+    behave (reflection and refraction, for example).
+
+
 * :func:`~fatiando.seismic.wavefd.elastic_psv`: Simulates the coupled P and SV
   elastic waves using the Parsimonious Staggered Grid method of Luo and
   Schuster (1990)
@@ -125,6 +134,7 @@ finite-differencing of the wave equation, Geophysical Research Letters, 17(2),
 
 """
 from __future__ import division
+import warnings
 
 import numpy
 import scipy.sparse
@@ -144,6 +154,12 @@ except:
     _nonreflexive_psv_boundary_conditions = not_implemented
     _step_scalar = not_implemented
     _reflexive_scalar_boundary_conditions = not_implemented
+
+
+# Tell users at import time that this code is not very trustworthy
+warnings.warn("This code is experimental and poorly tested! " +
+              "We cannot guarantee that the results are accurate and " +
+              "strongly discourage use of this code for research purposes.")
 
 
 class MexHatSource(object):